dom/tests/mochitest/geolocation/test_shutdown.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=716127
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=716127">Mozilla Bug 716127</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

function successCallback(position) {
  check_geolocation(position);

  SpecialPowers.pushPrefEnv({'set': [['geo.timeout', 100]]}, function() {
      delay_geolocationProvider(1000, function() {
        force_prompt(true, function() {
          navigator.geolocation.getCurrentPosition(success2, handle_error, {maximumAge: 0});
        });
      });

  });
}

function errorCallback() {
  ok(false, "unexpected error");
  SimpleTest.finish();
}

function test1() {
  navigator.geolocation.getCurrentPosition(successCallback, errorCallback);
}

function success2(position) {
  check_geolocation(position);
  SimpleTest.finish();
}

function handle_error() {
  ok(false, "geolocation provider should not have timed out");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
